Blog App 4.21-4.23

+++ Part4/blog-list/controllers/blogs.js
@@ -1,9 +1,5 @@
-const jwt = require('jsonwebtoken')
-const config = require('../utils/config')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
-
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
@@ -12,14 +8,7 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
-  const decodedToken = jwt.verify(request.token, config.SECRET)
-  if (!decodedToken.id) {
-    return response.status(401).json({ error: 'token invalid' })
-  }
-  const user = await User.findById(decodedToken.id)
-  if (!user) {
-    return response.status(400).json({ error: 'userId missing or not valid' })
-  }
+  const user = request.user
 
   const blog = new Blog({
     title: body.title,
@@ -38,7 +27,19 @@ blogsRouter.post('/', async (request, response) => {
 
 blogsRouter.delete('/:id', async (request, response) => {
   const id = request.params.id
+  const user = request.user
+
+  const blog = await Blog.findById(id)
+  if(!blog) {
+    return response.status(204).end()
+  }
+  if ( blog.user.toString() !== user._id.toString() ) {
+    return response.status(401).json({ error: 'user invalid' })
+  }
+
   await Blog.findByIdAndDelete(id)
+  user.blogs = user.blogs.filter(userBlog => userBlog.toString() !== id)
+  await user.save()
   response.status(204).end()
 })
 
@@ -52,6 +53,10 @@ blogsRouter.put('/:id', async (request, response) => {
   const blog = await Blog.findById(id)
   if (!blog) {
     return response.status(404).end()
+  }
+  const user = request.user
+  if ( blog.user.toString() !== user._id.toString() ) {
+    return response.status(401).json({ error: 'user invalid' })
   }
 
   // Update blog likes
